# Improve memory usage and overall performance