## What's in this PR?

This builds on the changes in #3 by applying the memory usage / allocation / performance optimizations.

### Changes to FF1:

1. Use a common backing array buffer for holding the variable sized `Q`, `PQ`, `Y`, `R`, `xored` buffers. `P` remains a single `blockSize` array on the stack. Example for reference on long inputs: ![image](https://user-images.githubusercontent.com/4002512/29528478-317f8e14-866a-11e7-95ec-0d364d2cd641.png)
2. `numA`, `numB`, and `numC` are swapped via assignment and `SetBytes` rather than repetitive `SetString` and `Text` calls inside the Feistel round for loop
3. Pre-calculate the moduli based on `m=u` (i is even) and `m=v` (i is odd) since there are only 2 that are used in an alternating fashion. 
4. Added some test cases/benchmarks for long inputs
5. General cleanup, some manual function inlining 

On average, FF1 performance improves quite a bit from the previous `master` HEAD (3d2762d3b590b2976bc536f5c105f60ab1eb061b )

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkEncrypt/Sample7-8     20627         5172          -74.93%

benchmark                      old allocs     new allocs     delta
BenchmarkEncrypt/Sample7-8     299            43             -85.62%

benchmark                      old bytes     new bytes     delta
BenchmarkEncrypt/Sample7-8     11248         1424          -87.34%
```

### Changes to FF3:

1. Utility function cleanup
    1. Manually inlined some function logic and removed the redundant functions. 
    2. `rev` now uses `revB` which `string`/`[]byte` conversions. 
    3. remove `cbcEncryptor` since input blocks are always 16 bytes with iv=0, so just use normal AES (ECB)
    4. `ciph` has been replaced with direct calls to aes encrypt
2. General cleanup, some similar changes from FF1

### MINOR BREAKING CHANGE TO BOTH:

`ff1` and `ff3` NewCipher now return a `Cipher` rather than `*Cipher` and methods are now on value receivers instead of pointer receivers for consistency. Because of Go's transparent indirection and since the struct fields are not exported, **this shouldn't cause issues for most practical uses**, as the only way to interact with the cipher is using `Encrypt` and `Decrypt`, which remain unchanged. 

This was done because in situations where many `NewCipher` calls must be made (ex. tweak is different for each piece of data), then using value semantics causes less allocations. 

## TODOs

  - [x] Apply same to FF1 Decrypt
  - [x] Apply same to FF3 Encrypt and Decrypt