@anitgandhi 2017-08-04
Preliminary benchmark improvements

```

benchmark                      old ns/op     new ns/op     delta
BenchmarkEncrypt/Sample1-8     19549         10515         -46.21%
BenchmarkEncrypt/Sample2-8     18519         10811         -41.62%
BenchmarkEncrypt/Sample3-8     21497         13406         -37.64%
BenchmarkEncrypt/Sample4-8     19548         10321         -47.20%
BenchmarkEncrypt/Sample5-8     18363         10577         -42.40%
BenchmarkEncrypt/Sample6-8     21806         13281         -39.09%
BenchmarkEncrypt/Sample7-8     19591         10600         -45.89%
BenchmarkEncrypt/Sample8-8     18453         10507         -43.06%
BenchmarkEncrypt/Sample9-8     21819         13529         -37.99%

benchmark                      old allocs     new allocs     delta
BenchmarkEncrypt/Sample1-8     301            100            -66.78%
BenchmarkEncrypt/Sample2-8     296            103            -65.20%
BenchmarkEncrypt/Sample3-8     289            101            -65.05%
BenchmarkEncrypt/Sample4-8     303            100            -67.00%
BenchmarkEncrypt/Sample5-8     296            103            -65.20%
BenchmarkEncrypt/Sample6-8     289            101            -65.05%
BenchmarkEncrypt/Sample7-8     299            100            -66.56%
BenchmarkEncrypt/Sample8-8     292            100            -65.75%
BenchmarkEncrypt/Sample9-8     289            101            -65.05%

benchmark                      old bytes     new bytes     delta
BenchmarkEncrypt/Sample1-8     11264         3040          -73.01%
BenchmarkEncrypt/Sample2-8     10976         3040          -72.30%
BenchmarkEncrypt/Sample3-8     11584         3408          -70.58%
BenchmarkEncrypt/Sample4-8     11280         3040          -73.05%
BenchmarkEncrypt/Sample5-8     10976         3040          -72.30%
BenchmarkEncrypt/Sample6-8     11584         3408          -70.58%
BenchmarkEncrypt/Sample7-8     11248         3040          -72.97%
BenchmarkEncrypt/Sample8-8     10960         3040          -72.26%
BenchmarkEncrypt/Sample9-8     11584         3408          -70.58%
```